bot/requirements.txt
python-telegram-bot==21.10
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
pytest==8.3.4
pytest-asyncio==0.25.0
respx==0.22.0
